requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
supabase==2.11.0
openai==1.58.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
openpyxl==3.1.5
pydantic==2.10.4
httpx==0.28.1
